split declaration+implementation
